Do not "force" the QC stages (#263)
* Enable QC reporting to Slack. Minor touches to Slack messages.

* Add readme

* Fix test

* Fix threshold for PCT_PF_READS_ALIGNED

* Remove backticks (breaking inspect)

* More threshold fix

* Add send_message function into slack module

* Remove slack creds from large-cohort config for now (token secret is available only for seqr project at the moment, would need to track it with pulumi to make it accessible by all datasets)

* Do no force QC stages anymore
We do not need to "force" the QC stages (MutliQC and SomalierPedigree) as non-html outputs are now parametrised by the samples hash (i.e. `h = dataset.alignment_inputs_hash()`), like the MultiQC JSON and the Somliaer TSV files, that would force the stage rerun if the set of samples has changed.