Use hard-coded in-repo test data for test/test_large_cohort.py
PR #326 changed `predetermined_qc_variants` so that its path is read
from an external config file. However for these unit tests, there is
no config other than what is hard-coded in the `TOML` string.

Revert (most of) the previous PR so that this path is hard-coded
(so refers to test/data/large_cohort/.../pre_ld_pruning_qc_variants.ht
within this repository) just as the other paths in `TOML` are.