Use current GitHub Action versions and set test timeout (PR #354)
Prevent "Node.js 12 actions are deprecated" warnings by updating to
the current releases of Actions used. Use the current releases instead
of `@main` too, as we don't want to be affected by bleeding-edge bugs.

Set a 20 minute timeout instead of the default six hours. The slowest
test currently takes ~5 minutes, but seqr_loader_dry has recently been
affected by an infinite loop in hail -- so we'd prefer not to wait for
several hours to see it time out.