Represent overwrite=True => non-optional Job in type annotations
When overwrite=True, can_reuse() is always false and hence markdup()
will always return a Job rather than None. Add overloads so that this
is represented in the type annotations for `overwrite: Literal[True]`.

Take advantage of this in mito.py to remove unnecessary asserts.
(Otherwise these asserts ought to be before the `jobs.append(...)` lines.)