Handle column inconsistencies when unioning background datasets
For data generated at different times, the tables being unioned
in add_background() may have slightly differing sets of columns.
For example, we've encountered a case where one table has an
`autosomal_mean_dp` column but other(s) do not. Use unify=True
to allow the code to continue despite these minor extra columns,
but require this to be turned on in the config only when necessary.