[Quick Fix] Multiple Active Analysis Inputs  (#1187)
* Pull gvcf or cram objects based on a workflow param
This enables prod pipes to determine which inputs should be used when there are multiple active analyses for an sg. This PR pulls a param from config, and feeds that to the metamist query to filter analyses accordingly

* Create the stage name attribute for a SequencingGroup
This allows for later reference r.e. which gvcf or which cram was used. The intention is to pass this onto downstream analyses meta for provenance

* Update todo comment to add further clarity
r.e. the purpose of the gvcf_stage_name and cram_stage_name properties i.e. speaks to the provenance of the file.

* set the config_retrieve default to None
This is an optional param, and should not break pipelines if it does not exist

* Store get_x_meta() methods as variable to reduce calls

* Move the filtering out of the metamist query
And instead have filter analyses logic after a general pull. This allows for more flexibility and less complex metamist queries

* take the cram source as input so analyses can be filtered on it

* Fix linter and reduce function calls
double quotes somehow missed and the function output is already stored as a variable so no need to recall it

* Switch meta call to stage to match metamist

---------

Co-authored-by: Michael Harper <michael.harper@populationgenomics.org.au>