Update our cpg_workflows entry in images.toml directly
Previously cpg_workflows was a special case in the images repo's
deployment workflows, which had explicit code to add an entry for
cpg_workflows. The images repo had a monopoly on updating the GCS
bucket images.toml file to avoid race conditions.

Due to the images revamp, the images repo no longer updates images.toml.
In the new world order, the only entry that we'll keep updating (for now)
is cpg_workflows, so it is now safe to do that here directly after we
build and upload the docker image.

Remove shell override so we use the default which includes `-e`,
so we avoid overwriting the cloud images.toml when things go wrong.